add: more app config
note: not fully tested all config combination

+++ game/src/main.cpp
@@ -8,7 +8,7 @@ int main() {
     config.title = "Axis Engine - Game";
     config.width = 1280;
     config.height = 720;
-    config.logLevel = static_cast<int>(LogLevel::Verbose); // Default log level
+    config.logLevel = LogLevel::Verbose; // Default log level
 
     if (app->Initialize(config)) {
         app->PushState<GameState>();
